argc: refactor: ast structure
