+++ argc/include/type_checker.hh
@@ -1,5 +1,5 @@
 #pragma once
-
+#include <map>
 #include "ast.hh"
 
 namespace ast {
