+++ argc/include/x86_64_codegen.hh
@@ -1,5 +1,6 @@
 #include <stack>
 #include "ast.hh"
+#include <sstream>
 
 namespace ast {
